# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediCare HMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      parameters:
        - { in: query, name: patientId, schema: { type: integer } }
        - { in: query, name: doctorId, schema: { type: integer } }
        - { in: query, name: department, schema: { type: string } }
        - { in: query, name: date, schema: { type: string } }
        - { in: query, name: status, schema: { type: string } }
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Appointment" }
    post:
      operationId: createAppointment
      tags: [appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppointmentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppointmentUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
    delete:
      operationId: cancelAppointment
      tags: [appointments]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: Cancelled }
...
  /beds:
    get:
      operationId: listBeds
      tags: [beds]
      responses:
        "200":
          description: Beds
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Bed" }
    post:
      operationId: createBed
      tags: [beds]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BedInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bed" }
  /beds/{id}/assign:
    post:
      operationId: assignBed
      tags: [beds]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BedAssignmentInput" }
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bed" }
  /beds/{id}/discharge:
    post:
      operationId: dischargeBed
      tags: [beds]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Discharged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bed" }
...
  /notifications/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendNotificationInput" }
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationLog" }
...
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, patientId, patientName, doctorId, doctorName, department, scheduledAt, status, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: string }
        doctorId: { type: integer }
        doctorName: { type: string }
        department: { type: string }
        scheduledAt: { type: string }
        status: { type: string }
        reason: { type: ["string", "null"] }
        tokenNumber: { type: ["integer", "null"] }
        createdAt: { type: string }
    AppointmentInput:
      type: object
      required: [patientId, doctorId, department, scheduledAt]
      properties:
        patientId: { type: integer }
        doctorId: { type: integer }
        department: { type: string }
        scheduledAt: { type: string }
        reason: { type: string }
    AppointmentUpdate:
      type: object
      properties:
        scheduledAt: { type: string }
        status: { type: string }
        reason: { type: string }
...
    Bed:
      type: object
      required: [id, code, ward, status, createdAt]
      properties:
        id: { type: integer }
        code: { type: string }
        ward: { type: string }
        status: { type: string }
        patientId: { type: ["integer", "null"] }
        patientName: { type: ["string", "null"] }
        admittedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    BedInput:
      type: object
      required: [code, ward]
      properties:
        code: { type: string }
        ward: { type: string }
        floor: { type: ["string", "null"] }
        bedType: { type: ["string", "null"] }
        genderPolicy: { type: ["string", "null"], enum: [male, female, any, null] }
        ageMinYears: { type: ["integer", "null"] }
        ageMaxYears: { type: ["integer", "null"] }
        dailyRate: { type: ["string", "null"] }
    BedAssignmentInput:
      type: object
      required: [patientId]
      properties:
        patientId: { type: integer }
...
    SendNotificationInput:
      type: object
      required: [eventKey, channel, patientId, variables]
      properties:
        eventKey: { type: string }
        channel: { type: string }
        patientId: { type: integer }
        recipientPhone: { type: string }
        variables:
          type: object
          additionalProperties: { type: string }
    NotificationLog:
      type: object
      required: [id, eventKey, channel, status, renderedBody, recipientPhone, sentAt]
      properties:
        id: { type: integer }
        patientId: { type: ["integer", "null"] }
        patientName: { type: ["string", "null"] }
        eventKey: { type: string }
        channel: { type: string }
        templateId: { type: ["integer", "null"] }
        renderedBody: { type: string }
        recipientPhone: { type: string }
        status: { type: string }
        providerRef: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        sentAt: { type: string }
...